+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "attoworld"
-version = "0.1.0"
-edition = "2021"
+version = "0.2.0"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
