+++ src/frog/mod.rs
@@ -1,2 +1,4 @@
 mod core;
 pub use core::{reconstruct_frog, FrogType};
+pub mod ptychographic;
+use ptychographic::*;
